refactor: extract the retry asking module
